fixed bug related to JobFinishedAck
git-svn-id: https://svn.itwm.fhg.de/svn/SDPA@1346 c3adf61f-8968-0410-94d0-83abfdf586cf

+++ trunk/sdpa/branches/rotaru/sdpa/daemon/Scheduler.hpp
@@ -15,7 +15,7 @@ namespace daemon {
 
 	 virtual Worker::ptr_t& findWorker(const Worker::worker_id_t&  ) throw(WorkerNotFoundException) = 0;
 	 virtual void addWorker(const  Worker::ptr_t& ) = 0;
-	 virtual int  numberOfWorkers() = 0;
+	 virtual size_t  numberOfWorkers() = 0;
 	 virtual void schedule(Job::ptr_t& pJob) = 0;
 	 virtual void schedule(gwes::activity_t& act)=0;
 	 virtual void start()=0;
