+++ trunk/sdpa/branches/rotaru/tests/sdpa/SchedulerTestImpl.hpp
@@ -22,7 +22,7 @@ class SchedulerTestImpl : public Scheduler {
     virtual void schedule(gwes::activity_t& act) {};
     Worker::ptr_t& findWorker(const Worker::worker_id_t&  ) throw(WorkerNotFoundException);
     void addWorker(const  Worker::ptr_t& );
-    int numberOfWorkers() { return 1; }
+    size_t numberOfWorkers() { return 1; }
 
     // thread related functions
     void start();
